Separate derived state from UI helpers
