+++ composeApp/src/jvmMain/kotlin/io/github/hayatoyagi/prvisualizer/navigation/NavigationManager.kt
@@ -1,9 +1,9 @@
 package io.github.hayatoyagi.prvisualizer.navigation
 
+import io.github.hayatoyagi.prvisualizer.filetree.parentPathOf
 import io.github.hayatoyagi.prvisualizer.state.NavigationState
 import io.github.hayatoyagi.prvisualizer.state.resetNavigation
 import io.github.hayatoyagi.prvisualizer.state.resetViewport
-import io.github.hayatoyagi.prvisualizer.ui.shared.parentPathOf
 
 /**
  * Manages navigation history, viewport, and directory focus.
