+++ composeApp/src/jvmMain/kotlin/io/github/hayatoyagi/prvisualizer/state/VisualizerState.kt
@@ -5,13 +5,10 @@ import androidx.compose.ui.state.ToggleableState
 import io.github.hayatoyagi.prvisualizer.AppError
 import io.github.hayatoyagi.prvisualizer.FileCommit
 import io.github.hayatoyagi.prvisualizer.PullRequest
+import io.github.hayatoyagi.prvisualizer.filetree.collectAllDirectories
+import io.github.hayatoyagi.prvisualizer.filetree.collectAllFiles
+import io.github.hayatoyagi.prvisualizer.filetree.findDirectory
 import io.github.hayatoyagi.prvisualizer.github.GitHubSnapshot
-import io.github.hayatoyagi.prvisualizer.ui.prlist.filterPrs
-import io.github.hayatoyagi.prvisualizer.ui.shared.collectAllDirectories
-import io.github.hayatoyagi.prvisualizer.ui.shared.collectAllFiles
-import io.github.hayatoyagi.prvisualizer.ui.shared.computeDirectoryOverlayByPath
-import io.github.hayatoyagi.prvisualizer.ui.shared.computeFileOverlayByPath
-import io.github.hayatoyagi.prvisualizer.ui.shared.findDirectory
 
 sealed interface RepoSelectionState {
     data object Idle : RepoSelectionState
