+++ composeApp/src/jvmMain/kotlin/io/github/hayatoyagi/prvisualizer/ui/dialog/DialogHost.kt
@@ -7,6 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.graphics.Color
 import io.github.hayatoyagi.prvisualizer.AppError
+import io.github.hayatoyagi.prvisualizer.filetree.findFileNode
 import io.github.hayatoyagi.prvisualizer.repository.RepoState
 import io.github.hayatoyagi.prvisualizer.state.DialogState
 import io.github.hayatoyagi.prvisualizer.state.RepoSelectionState
@@ -14,7 +15,6 @@ import io.github.hayatoyagi.prvisualizer.state.SnapshotFetchState
 import io.github.hayatoyagi.prvisualizer.ui.file.FileDetailsDialog
 import io.github.hayatoyagi.prvisualizer.ui.prlist.PrDetailsDialog
 import io.github.hayatoyagi.prvisualizer.ui.repo.RepoPickerDialog
-import io.github.hayatoyagi.prvisualizer.ui.shared.findFileNode
 import io.github.hayatoyagi.prvisualizer.ui.theme.AppColors
 
 /**
