--- composeApp/src/jvmMain/kotlin/io/github/hayatoyagi/prvisualizer/ui/shared/UiModels.kt
@@ -1,16 +1,0 @@
-package io.github.hayatoyagi.prvisualizer.ui.shared
-
-import io.github.hayatoyagi.prvisualizer.ChangeType
-import io.github.hayatoyagi.prvisualizer.PullRequest
-
-data class FileOverlay(
-    val prs: List<PullRequest>,
-    val dominantType: ChangeType,
-    val density: Float,
-)
-
-data class DirectoryOverlay(
-    val prs: List<PullRequest>,
-    val dominantType: ChangeType?,
-    val density: Float,
-)
