+++ composeApp/src/jvmMain/kotlin/io/github/hayatoyagi/prvisualizer/ui/treemap/TreemapPane.kt
@@ -15,8 +15,8 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.IntSize
 import io.github.hayatoyagi.prvisualizer.FileNode
 import io.github.hayatoyagi.prvisualizer.TreemapNode
-import io.github.hayatoyagi.prvisualizer.ui.shared.DirectoryOverlay
-import io.github.hayatoyagi.prvisualizer.ui.shared.FileOverlay
+import io.github.hayatoyagi.prvisualizer.state.DirectoryOverlay
+import io.github.hayatoyagi.prvisualizer.state.FileOverlay
 import io.github.hayatoyagi.prvisualizer.ui.treemap.components.TreemapViewport
 import io.github.hayatoyagi.prvisualizer.ui.treemap.handlers.INITIAL_ZOOM
 import io.github.hayatoyagi.prvisualizer.ui.treemap.handlers.MAX_ZOOM
